Preparing for breaking API changes
